Add NestedObject class to allow for custom structures.
Previously, `tree` was only able to handle built-in
types and structures which implemented the standard
library's `Mapping` or `Sequence` interfaces.
However, there are cases where a class might not quite
match the interface of a sequence or map despite being
and index-able composite structure.

NestedObject defines an interface through which such
types can be compatible with `tree`. It does so
through protected methods so as not to disturb the
public API of such types.

PiperOrigin-RevId: 316751166